CodeReview: Seed param moved to builder
