+++ pallets/asset/src/benchmarking.rs
@@ -33,7 +33,6 @@ use sp_std::{
     prelude::*,
 };
 
-const SEED: u32 = 0;
 const MAX_DOCS_PER_ASSET: u32 = 64;
 const MAX_DOC_URI: usize = 4096;
 const MAX_DOC_NAME: usize = 1024;
@@ -109,7 +108,7 @@ fn make_classic_ticker<T: Trait>(eth_owner: ethereum::EthereumAddress, ticker: T
 
 fn make_extension<T: Trait>(is_archive: bool) -> SmartExtension<T::AccountId> {
     // Simulate that extension was added.
-    let extension_id = UserBuilder::<T>::default().build("extension", 0).account;
+    let extension_id = UserBuilder::<T>::default().build("extension").account;
     let extension_details = SmartExtension {
         extension_type: SmartExtensionType::TransferManager,
         extension_name: b"PTM".into(),
@@ -139,15 +138,15 @@ benchmarks! {
             registration_length: Some((60 * 24 * 60 * 60).into()),
         });
 
-        let caller = UserBuilder::<T>::default().generate_did().build("caller", SEED);
+        let caller = UserBuilder::<T>::default().generate_did().build("caller");
         // Generate a ticker of length `t`.
         let ticker = Ticker::try_from(vec![b'A'; TICKER_LEN].as_slice()).unwrap(); }: _(caller.origin, ticker.clone()) verify {
         assert_eq!(Module::<T>::is_ticker_available(&ticker), false);
     }
 
     accept_ticker_transfer {
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
-        let new_owner = UserBuilder::<T>::default().generate_did().build("new_owner", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
+        let new_owner = UserBuilder::<T>::default().generate_did().build("new_owner");
         let ticker = make_ticker::<T>(owner.origin().into());
 
         Module::<T>::asset_ownership_relation(owner.did(), ticker.clone());
@@ -165,8 +164,8 @@ benchmarks! {
     }
 
     accept_asset_ownership_transfer {
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
-        let new_owner = UserBuilder::<T>::default().generate_did().build("new_owner", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
+        let new_owner = UserBuilder::<T>::default().generate_did().build("new_owner");
 
         let ticker = make_asset::<T>(&owner);
 
@@ -207,7 +206,7 @@ benchmarks! {
         let identifiers: Vec<AssetIdentifier> =
             iter::repeat(AssetIdentifier::cusip(*b"17275R102").unwrap()).take(i as usize).collect();
         let fundr = FundingRoundName::from(vec![b'F'; f as usize].as_slice());
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
         let total_supply: T::Balance = 1_000_000.into();
 
         let token = SecurityToken {
@@ -228,7 +227,7 @@ benchmarks! {
 
 
     freeze {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
     }: _(owner.origin, ticker.clone())
     verify {
@@ -236,7 +235,7 @@ benchmarks! {
     }
 
     unfreeze {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
 
         Module::<T>::freeze( owner.origin().into(), ticker.clone())
@@ -253,7 +252,7 @@ benchmarks! {
         let n in 1 .. T::AssetNameMaxLength::get() as u32;
 
         let new_name = AssetName::from(vec![b'N'; n as usize].as_slice());
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
     }: _(owner.origin(), ticker.clone(), new_name.clone())
     verify {
@@ -262,7 +261,7 @@ benchmarks! {
     }
 
     issue {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
 
     }: _(owner.origin, ticker.clone(), 1_000_000.into())
@@ -273,7 +272,7 @@ benchmarks! {
 
 
     redeem {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
     }: _(owner.origin, ticker.clone(), 600_000.into())
     verify {
@@ -282,7 +281,7 @@ benchmarks! {
     }
 
     make_divisible {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_indivisible_asset::<T>(&owner);
     }: _(owner.origin, ticker)
     verify {
@@ -294,7 +293,7 @@ benchmarks! {
         // It starts at 1 in order to get something for `verify` section.
         let d in 1 .. MAX_DOCS_PER_ASSET;
 
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
         let docs = iter::repeat( make_document()).take( d as usize).collect::<Vec<_>>();
 
@@ -308,7 +307,7 @@ benchmarks! {
     remove_documents {
         let d in 1 .. MAX_DOCS_PER_ASSET;
 
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
         let docs = iter::repeat( make_document()).take( MAX_DOCS_PER_ASSET as usize).collect::<Vec<_>>();
         Module::<T>::add_documents( owner.origin().into(), docs.clone(), ticker)
@@ -326,7 +325,7 @@ benchmarks! {
      set_funding_round {
         let f in 1 .. T::FundingRoundNameMaxLength::get() as u32;
 
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
 
         let fundr = FundingRoundName::from(vec![b'X'; f as usize].as_slice());
@@ -339,7 +338,7 @@ benchmarks! {
      update_identifiers {
          let i in 1 .. MAX_IDENTIFIERS_PER_ASSET;
 
-         let owner = UserBuilder::default().generate_did().build("owner", SEED);
+         let owner = UserBuilder::default().generate_did().build("owner");
          let ticker = make_asset::<T>(&owner);
 
          let identifiers: Vec<AssetIdentifier> =
@@ -351,7 +350,7 @@ benchmarks! {
      }
 
      add_extension {
-         let owner = UserBuilder::default().generate_did().build("owner", SEED);
+         let owner = UserBuilder::default().generate_did().build("owner");
          let ticker = make_asset::<T>(&owner);
          let ext_details = make_extension::<T>(false);
          let ext_id = ext_details.extension_id.clone();
@@ -361,7 +360,7 @@ benchmarks! {
      }
 
      archive_extension {
-         let owner = UserBuilder::default().generate_did().build("owner", SEED);
+         let owner = UserBuilder::default().generate_did().build("owner");
          let ticker = make_asset::<T>(&owner);
          let ext_details = make_extension::<T>(false);
          let ext_id = ext_details.extension_id.clone();
@@ -375,7 +374,7 @@ benchmarks! {
      }
 
      unarchive_extension {
-         let owner = UserBuilder::default().generate_did().build("owner", SEED);
+         let owner = UserBuilder::default().generate_did().build("owner");
          let ticker = make_asset::<T>(&owner);
          let ext_details = make_extension::<T>(true);
          let ext_id = ext_details.extension_id.clone();
@@ -389,7 +388,7 @@ benchmarks! {
      }
 
      remove_smart_extension {
-         let owner = UserBuilder::default().generate_did().build("owner", SEED);
+         let owner = UserBuilder::default().generate_did().build("owner");
          let ticker = make_asset::<T>(&owner);
          let ext_details = make_extension::<T>(false);
          let ext_id = ext_details.extension_id.clone();
@@ -401,7 +400,7 @@ benchmarks! {
      }
 
     remove_primary_issuance_agent {
-        let owner = UserBuilder::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::default().generate_did().build("owner");
         let ticker = make_asset::<T>(&owner);
     }: _(owner.origin, ticker.clone())
     verify {
@@ -411,7 +410,7 @@ benchmarks! {
 
 
     claim_classic_ticker {
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
         let owner_eth_sk = secp256k1::SecretKey::parse(&keccak_256(b"owner")).unwrap();
         let owner_eth_pk = ethereum::address(&owner_eth_sk);
 
@@ -426,7 +425,7 @@ benchmarks! {
     }
 
     reserve_classic_ticker {
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
 
         let ticker :Ticker = b"ACME"[..].try_into().unwrap();
         let config = make_default_reg_config::<T>();
@@ -442,8 +441,8 @@ benchmarks! {
     }
 
     accept_primary_issuance_agent_transfer {
-        let owner = UserBuilder::<T>::default().generate_did().build("owner", SEED);
-        let primary_issuance_agent = UserBuilder::<T>::default().generate_did().build("1stIssuance", SEED);
+        let owner = UserBuilder::<T>::default().generate_did().build("owner");
+        let primary_issuance_agent = UserBuilder::<T>::default().generate_did().build("1stIssuance");
         let ticker = make_asset::<T>(&owner);
 
         let auth_id = identity::Module::<T>::add_auth(
