+++ pallets/pips/src/benchmarking.rs
@@ -63,7 +63,8 @@ fn make_voters<T: Trait>(
                 ..
             } = UserBuilder::<T>::default()
                 .generate_did()
-                .build(prefix, i as u32);
+                .seed(i as u32)
+                .build(prefix);
             (account, origin, did.unwrap())
         })
         .collect()
@@ -94,9 +95,7 @@ fn pips_and_votes_setup<T: Trait>(
     };
     let hi_voters = make_voters::<T>(voters_a_num, "hi");
     let bye_voters = make_voters::<T>(voters_b_num, "bye");
-    let User { origin, did, .. } = UserBuilder::<T>::default()
-        .generate_did()
-        .build("initial", 0);
+    let User { origin, did, .. } = UserBuilder::<T>::default().generate_did().build("initial");
     let did = did.expect("no did in pips_and_votes_setup");
     for i in 0..PROPOSALS_NUM {
         let (proposal, url, description) = make_proposal::<T>();
@@ -192,7 +191,7 @@ benchmarks! {
 
     propose_from_community {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         identity::CurrentDid::put(did.unwrap());
         let (proposal, url, description) = make_proposal::<T>();
     }: propose(origin, proposal, 42.into(), Some(url.clone()), Some(description.clone()))
@@ -206,7 +205,7 @@ benchmarks! {
     // `propose` from a committee origin.
     propose_from_committee {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         identity::CurrentDid::put(did.unwrap());
         let (proposal, url, description) = make_proposal::<T>();
         let origin = T::UpgradeCommitteeVMO::successful_origin();
@@ -224,7 +223,7 @@ benchmarks! {
 
     vote {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         identity::CurrentDid::put(did.unwrap());
         let (proposal, url, description) = make_proposal::<T>();
         Module::<T>::propose(
@@ -241,7 +240,7 @@ benchmarks! {
         cast_votes::<T>(0, nay_voters.as_slice(), false)?;
         // Cast an opposite vote.
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("voter", 0);
+            UserBuilder::<T>::default().generate_did().build("voter");
         identity::CurrentDid::put(did.unwrap());
         let voter_deposit = 43.into();
         // Cast an opposite vote.
@@ -270,7 +269,7 @@ benchmarks! {
 
     reject_proposal {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         identity::CurrentDid::put(did.unwrap());
         let (proposal, url, description) = make_proposal::<T>();
         let deposit = 42.into();
@@ -293,7 +292,7 @@ benchmarks! {
 
     prune_proposal {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         identity::CurrentDid::put(did.unwrap());
         let (proposal, url, description) = make_proposal::<T>();
         Module::<T>::propose(
@@ -318,7 +317,7 @@ benchmarks! {
 
     reschedule_execution {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         let did = did.expect("missing did in reschedule_execution");
         identity::CurrentDid::put(did);
         let (proposal, url, description) = make_proposal::<T>();
@@ -342,7 +341,7 @@ benchmarks! {
 
     clear_snapshot {
         let User { account, origin, did, .. } =
-            UserBuilder::<T>::default().generate_did().build("proposer", 0);
+            UserBuilder::<T>::default().generate_did().build("proposer");
         let did = did.expect("missing did in clear_snapshot");
         identity::CurrentDid::put(did);
         let (proposal, url, description) = make_proposal::<T>();
